Move GGA selector to stream
Rename the receiver packet input to Packet so gpsevent can feed successfully processed packets through a narrow selector interface while the selector handles GGA filtering.
